[DSCP-248] Shutdown workers and listeners gracefully
Wait for workers and listeners to terminate when user does ^C.
I hoped this will help me to see stacktraces from workers, but got kek instead.

This also means that if someone grabs a lock for too long, appplication won't
terminate until lock is released (because internals are masked uninteruptibly).