[DSCP-303] Replace chmod with error message
When reading a key file from disk, check for permissions and throw an
error instead of blindly chmodding the file.

When we did not create the file ourselves, we can not be sure we have
the necessary permissions to chmod the file, which can cause issues in
sandboxed environments such as systemd with DynamicUser=true.