[DSCP-314] Fix CLI parser of `CommitteeSecret`
Originally that field was optional, altering kind of committee used depending
on whether the option was present or not.

In recent PR the CLI parser was modified so that it always expected
committee secret, and says there is no committee if it failed to parse (i.e.
it was empty).

This fix makes committee secret truly optional, error is raised if
empty committee secret was passed. Note that this does not make CLI parser
override config by default because still there is mandatory `comm-n` parameter.