[DSCP-390] Mock slotting (WIP)
Things are so that some actions are allowed only in testing (managing slot id manually)
or only in real mode (blocking until next slot). Ideally, this means that we need sligtly
different portion of constraints for producation and testing code.

In this commit I'm trying to implement that in what I see the least boilerplate-y way.