[DSCP-457] Bump `async` version to `2.2.1`
The previous version (2.1.1.1) of async (which we used) contains buggy `concurrently`.
Its documentation states that it is equivalent to combination of `withAsync` and `waitBoth`
but actually it uses a more smart implementation via `MVar`s and its behaviour differs:
`concurrently` does not wait for its children to terminate before call completes.

This may be dangerous, e.g. when used to allocate workers, later on termination they are
not awaited and the main thread may terminate earlier than cleanup occurs.